backend/requirements.txt
fastapi[all]==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy[asyncio]==2.0.35
asyncpg==0.30.0
pydantic==2.9.2
pydantic-settings==2.5.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
openai>=1.30.0
httpx>=0.27.0
chromadb==0.5.5
azure-cognitiveservices-speech==1.37.0
crewai==0.76.9
loguru==0.7.2
pymupdf==1.25.3
reportlab



